Most important change - added an option to create a DBAdaptor with a dna database handle.  All the dna fetching is done through this extra dna handle.  This allows us to create new databases with fresh features and genes without having to use up 4Gb worth of disk space.  All deleting from the dna table is disallowed.  Arne will probably shout at me for not using a proper adaptor.
Added percent_id,p_value,phase to supporting evidence in ExonAdaptor

Removed debug from ExternalWrapper

Gene_Obj - writes phase,p_value,percent_id from supporting features

Raw_Contig :
	   removed call to got_overlaps
	   removed calls to get_left_overlap (and right)
	   removed load_overlaps
	   removed unnecessary joiun to dna table
	   removed string checking (slow)
	   remove feature_file method
	   removed get_extra_features method
	   removed analysis method
	   removed filter_features method
	   all dna is fetched through the dna adaptor and
           no deleting is allowed from the dna table.

Utils:
	refixed phase problem

CloneAdaptor

        removed overlaps
        added warning about deleting from dna table if using
        a remote dna database

+++ modules/Bio/EnsEMBL/DBSQL/CloneAdaptor.pm
@@ -239,6 +239,7 @@ sub delete_by_dbID {
    my @dnas;
 
    # get a list of contigs to zap
+
    my $sth = $self->prepare("select internal_id,dna from contig where clone = $internal_id");
    my $res = $sth->execute;
 
@@ -255,11 +256,15 @@ sub delete_by_dbID {
    }
 
 
-   foreach my $dna (@dnas) {
+   if ($self->db ne $self->db->dnadb) {
+     $self->warn("Using a remote dna database - can't delete dna\n");
+   } else {
+     
+     foreach my $dna (@dnas) {
        $sth = $self->prepare("delete from dna where id = $dna");
        $res = $sth->execute;
-
-       # Mysql does not optimise or statements in where clauses
+       
+     }
 
    }
 
@@ -449,6 +454,7 @@ sub create_tables {
   } );
   $sth->execute();
 }
+
 
 
 1;
